Update parent in patches config to avoid a rebase.
In order to allow release-time patches to be ported to new release
tracks, this script creates new refs based on the source rosdistro name
with the intent that any patches applied in the source track will also
be applied in the destination track. This wasn't working because bloom
keeps track of the parent ref in its internal patch config file and will
"rebase" a branch if its parent and the current source branch don't
match.